girepository: Use standard types instead of glib specific

See merge request GNOME/glib!3780

+++ girepository/giarginfo.h
@@ -69,10 +69,10 @@ GI_AVAILABLE_IN_ALL
 GIScopeType            gi_arg_info_get_scope              (GIArgInfo *info);
 
 GI_AVAILABLE_IN_ALL
-gint                   gi_arg_info_get_closure_index      (GIArgInfo *info);
+gssize                 gi_arg_info_get_closure_index      (GIArgInfo *info);
 
 GI_AVAILABLE_IN_ALL
-gint                   gi_arg_info_get_destroy_index      (GIArgInfo *info);
+gssize                 gi_arg_info_get_destroy_index      (GIArgInfo *info);
 
 GI_AVAILABLE_IN_ALL
 GITypeInfo *           gi_arg_info_get_type_info          (GIArgInfo *info);
